moved deleteLine and selectedLineBlocks to delete-line.js
